+++ src/features/basic/prun-bugs.ts
@@ -1,5 +1,4 @@
 import { getPrunCssStylesheets } from '@src/infrastructure/prun-ui/prun-css';
-import { changeInputValue } from '@src/util';
 import $style from './prun-bugs.module.css';
 
 function removeMobileCssRules() {
@@ -32,27 +31,9 @@ function fixZOrder() {
   applyCssRule(`.${C.ScrollView.track}`, $style.scrollTrack);
 }
 
-function fixContractConditionEditor() {
-  // Condition editor fails to save the condition unless the user interacts with the input.
-  // Here we simulate this interaction.
-  tiles.observe('CONTD', tile => {
-    subscribe($$(tile.anchor, C.DraftConditionEditor.form), form => {
-      subscribe($$(form, 'input'), input => {
-        if (input.type !== 'text') {
-          return;
-        }
-        const value = input.value;
-        changeInputValue(input, '');
-        changeInputValue(input, value);
-      });
-    });
-  });
-}
-
 function init() {
   removeMobileCssRules();
   fixZOrder();
-  fixContractConditionEditor();
 
   // Prevents top-right user info from shrinking.
   applyCssRule(`.${C.Head.container}`, $style.head);